# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Dinosaur {
  id: number;
  name: string;
  description: string;
  period: string;
  diet: string;
  imageUrl?: string | null;
  likesCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDinosaur {
  name: string;
  description: string;
  period: string;
  diet: string;
  imageUrl?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type GetDinosaursParams = {
  search?: string;
};