# lib/api-zod/src/generated/types/dinosaur.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Dinosaur {
  id: number;
  name: string;
  description: string;
  period: string;
  diet: string;
  imageUrl?: string | null;
  likesCount: number;
  createdAt: Date;
  updatedAt: Date;
}